renamed the entity table
