+++ src/monitoring-new/system-state-snapshot.repository.ts
@@ -1,0 +1,5 @@
+import { EntityRepository, Repository } from 'typeorm';
+import { SystemStateSnapshot } from './system-state-snapshot.entity';
+
+@EntityRepository(SystemStateSnapshot)
+export class SystemStateSnapshotRepository extends Repository<SystemStateSnapshot> {}
