--- src/monitoring-new/system-state.repository.ts
@@ -1,5 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { SystemStateRecord } from './system-state.entity';
-
-@EntityRepository(SystemStateRecord)
-export class SystemStateRepository extends Repository<SystemStateRecord> {}
